Исправление замечаний
Исправила замечания:
1. Вынесла весь код в отдельный класс
2. Исправила объявление методов внутри другого метода
3. Исправила условие выполнения цикла while в методе Move()
4. Вынесла в переменные номера первой и последней ячейки поля
5. Исправила сообщение о неправильном указании номера ячейки в методе Move()
6. Добавила документацию к каждому методу
7. Вынесла "O" и "X" в константы класса
